+++ bin/upgrader.js
@@ -15,11 +15,11 @@ const E = module.exports = {};
 let crontab;
 
 E.start_upgrader = ()=>{
-    sudo_run(upgrader_cmd('--add-cron-job'));
+    sudo_run(upgrader_cmd('--add-cron-job'), null, {log: true});
 };
 
 E.stop_upgrader = ()=>{
-    sudo_run(upgrader_cmd('--remove-cron-job'));
+    sudo_run(upgrader_cmd('--remove-cron-job'), null, {log: true});
 };
 
 const upgrader_cmd = flag=>{
@@ -74,18 +74,19 @@ E.downgrade = cb=>{
 
 const bash_cmd = script=>'bash '+path.resolve(__dirname, script)+' /tmp/lpm_l';
 
-const sudo_run = (cmd, cb)=>{
-    const opt = {name: 'Upgrader LPM'};
+const sudo_run = (cmd, cb, opt={})=>{
     const exec = process.getuid()===0 ?
         child_process.exec.bind(child_process) :
         sudo_prompt.exec.bind(sudo_prompt);
-    exec(cmd, opt, (e, stdout, stderr)=>{
+    exec(cmd, {name: 'Upgrader LPM'}, (e, stdout, stderr)=>{
         if (cb)
             cb(e, stdout, stderr);
         if (e)
             return logger.error(e.message);
         if (stderr)
             logger.error(stderr);
+        if (opt.log && stdout)
+            logger.notice(stdout);
     });
 };
 
@@ -93,7 +94,7 @@ const add_cron_job = ()=>etask(function*(){
     const _this = this;
     const jobs = crontab.find({comment: JOB_NAME});
     if (jobs.length)
-        return console.log('CRON job already exists');
+        return console.log('Auto-upgrader is already installed and running');
     const node_path = process.argv[0];
     const script = path.resolve(__dirname, 'index.js');
     const set_path = `PATH=${process.env.PATH}`;
@@ -103,7 +104,13 @@ const add_cron_job = ()=>etask(function*(){
     crontab.create(cron_cmd, '*/15 * * * *', JOB_NAME);
     crontab.save(function(err){
         if (err)
-            console.error('could not save crontab: %s', err);
+            console.error('Could not save crontab: %s', err);
+        else
+        {
+            console.log('Auto-upgrader has been successfully installed');
+            console.log('To run LPM run \'luminati\' or \'luminati -d\' to'
+                +' keep it in the daemon mode');
+        }
         _this.continue();
     });
     yield _this.wait();
@@ -113,11 +120,13 @@ const remove_cron_job = ()=>etask(function*(){
     const _this = this;
     const jobs = crontab.find({comment: JOB_NAME});
     if (!jobs.length)
-        return console.log('CRON job does not exist');
+        return console.log('Auto-upgrader is not running. Nothing to stop');
     crontab.remove(jobs[0]);
     crontab.save(function(err){
         if (err)
-            console.error('could not delete job: %s', err);
+            console.error('Could not delete job: %s', err);
+        else
+            console.log('Auto-upgrader has been successfully stopped');
         _this.continue();
     });
     yield _this.wait();
@@ -125,10 +134,10 @@ const remove_cron_job = ()=>etask(function*(){
 
 const load_cron = ()=>etask(function*(){
     const _this = this;
-    require('crontab').load(function(err, _crontab){
+    require('lum_crontab').load(function(err, _crontab){
         if (err)
         {
-            console.error('could not load crontab: %s', err);
+            console.error('Could not load crontab: %s', err);
             return process.exit();
         }
         _this.continue(_crontab);
@@ -142,7 +151,7 @@ const main = ()=>etask(function*(){
     if (!argv.addCronJob && !argv.removeCronJob)
         return;
     if (process.getuid()!==0)
-        return console.error('you need to run the process with sudo');
+        return console.error('You need to run the process with sudo');
     crontab = yield load_cron();
     if (argv.addCronJob)
         return yield add_cron_job();