+++ test/common.js
@@ -103,7 +103,8 @@ E.http_proxy = port=>etask(function*(){
             _url = '127.0.0.1:'+proxy.https.address().port;
         }
         let req_port;
-        res.write(`HTTP/1.1 200 OK\r\nx-hola-ip: ${to_body(req).ip}\r\n\r\n`);
+        res.write(`HTTP/1.1 200 OK\r\nx-luminati-ip: ${to_body(req).ip}`
+            +'\r\n\r\n');
         if (req.method=='CONNECT')
             proxy.full_history.push(to_body(req));
         const socket = net.connect({